refactor: separate InputController
