+++ InputController/InputController.h
@@ -1,0 +1,19 @@
+//
+//  InputController.h
+//  InputController
+//
+//  Created by mi on 2024/12/1.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for InputController.
+FOUNDATION_EXPORT double InputControllerVersionNumber;
+
+//! Project version string for InputController.
+FOUNDATION_EXPORT const unsigned char InputControllerVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <InputController/PublicHeader.h>
+
+#import <rime_api_stdbool.h>
+#import <rime/key_table.h>
