+++ InputController/module.modulemap
@@ -1,0 +1,6 @@
+module InputController {
+  umbrella header "InputController.h"
+
+  export *
+  module * { export * }
+}
